Update server.js
Added a comment block containing a line that sets an additional saml.Strategy options property during the example uwshib.Strategy construction. This property defaults to 0 milliseconds, which assumes the client and the server are using the same authoritative time server. This addition allows for a small amount of skew between the client and the server's clock.